Simplify query, and resolve denominator issue overreporting rewards
The snapshots that are fetched for a given user for balances were not being fetched for 0 values
Thus a user who only participated in the last few hours of a campaign would be allocated their share for the entire elapsed time rather than just their participation time.

When doing initial testing queries hourly_balances with 0 values were being received. It is unclear why or how that was changed, or if it is inconsistent accross snapshots.

Changed to using hours elapsed using timestamps as the denominator.